// src/services/TH4/Tracuuthongtin/typing.d.ts
export interface DiplomaInfo {
    id: number;
    so_hieu_van_bang: string;
    so_vao_so: number;
    ma_sinh_vien: string;
    ho_ten: string;
    ngay_sinh: string;
    diem_trung_binh: number;
    xep_loai: string;
    he_dao_tao: string;
    noi_sinh: string;
    dan_toc: string;
    quyet_dinh_id: number;
}